[CODE REFACTOR: lua/LazyDeveloperHelper/python/* | NEW FILE/DIR: lua/LazyDeveloperHelper/python/logger.py, lua/LazyDeveloperHelper/python/c_installers/nuget_install.py] Divided logging message for file, refactored code by importing this file
