+++ lua/LazyDeveloperHelper/python/c_installers/nuget_install.py
@@ -1,0 +1,35 @@
+import shutil as sh
+import sys
+from subprocess import run, CalledProcessError
+
+from ..logger import log_message
+
+
+# --- CHECK NuGet is exist ---
+def nuget_exist() -> bool:
+    dotnet_path = sh.which("dotnet")
+    if dotnet_path:
+        log_message("Dotnet exists!", "info")
+        return True
+    else:
+        log_message("Dotnet isnt exists, try install it!", "critical")
+        return False
+
+
+# --- INSTALLLING LIBS ---
+def install_lib(lib_name: str):
+    message_for_run = ["dotnet", "add", "package", lib_name]
+    try:
+        result = run(message_for_run, check=True, capture_output=True, text=True)
+        log_message(f"NuGet install output:\n{result.stdout}", "info")
+    except CalledProcessError as err:
+        log_message(f"Dotnet install failed:\n{err.stderr}", "error")
+
+
+# --- POINT OF ENTER ---
+if __name__ == "__main__":
+    if not nuget_exist():
+        sys.exit(1)
+
+    for lib in sys.argv[1:]:
+        install_lib(lib)
