+++ lua/LazyDeveloperHelper/python/logger.py
@@ -1,0 +1,12 @@
+# -*- coding: utf-8 -*-
+
+
+# --- LOGGING MESSAGE ---
+def log_message(message: str, level: str = "info") -> None:
+    prefixes = {
+        "info": "\U0001f4cd",  # 📍
+        "success": "\U0001f4e6",  # 📦
+        "error": "\u274c",  # ❌
+    }
+
+    print(f"{prefixes.get(level, '\U0001f4cd')} {message}")
